Move controls to data package root
